Move RequestAPI to api module
