+++ ambry-replication/src/main/java/com.github.ambry.replication/ReplicationEngine.java
@@ -111,15 +111,11 @@ public abstract class ReplicationEngine implements ReplicationAPI {
   }
 
   /**
-   * Enables/disables replication of the given {@code ids} from {@code origins}. The disabling is in-memory and
-   * therefore is not valid across restarts.
-   * @param ids the {@link PartitionId}s to enable/disable it on.
-   * @param origins the list of datacenters from which replication should be enabled/disabled. Having an empty list
-   *                disables replication from all datacenters.
-   * @param enable whether to enable ({@code true}) or disable.
-   * @return {@code true} if disabling succeeded, {@code false} otherwise. Disabling fails if {@code origins} is empty
-   * or contains unrecognized datacenters.
+   * Starts up the replication engine.
+   * @throws ReplicationException
    */
+  public abstract void start() throws ReplicationException;
+
   @Override
   public boolean controlReplicationForPartitions(Collection<PartitionId> ids, List<String> origins, boolean enable) {
     if (origins.isEmpty()) {
@@ -136,13 +132,6 @@ public abstract class ReplicationEngine implements ReplicationAPI {
     return true;
   }
 
-  /**
-   * Updates the total bytes read by a remote replica from local store
-   * @param partitionId PartitionId to which the replica belongs to
-   * @param hostName HostName of the datanode where the replica belongs to
-   * @param replicaPath Replica Path of the replica interested in
-   * @param totalBytesRead Total bytes read by the replica
-   */
   @Override
   public void updateTotalBytesReadByRemoteReplica(PartitionId partitionId, String hostName, String replicaPath,
       long totalBytesRead) {
@@ -152,13 +141,6 @@ public abstract class ReplicationEngine implements ReplicationAPI {
     }
   }
 
-  /**
-   * Gets the replica lag of the remote replica with the local store
-   * @param partitionId The partition to which the remote replica belongs to
-   * @param hostName The hostname where the remote replica is present
-   * @param replicaPath The path of the remote replica on the host
-   * @return The lag in bytes that the remote replica is behind the local store
-   */
   @Override
   public long getRemoteReplicaLagFromLocalInBytes(PartitionId partitionId, String hostName, String replicaPath) {
     RemoteReplicaInfo remoteReplicaInfo = getRemoteReplicaInfo(partitionId, hostName, replicaPath);
@@ -196,11 +178,10 @@ public abstract class ReplicationEngine implements ReplicationAPI {
   }
 
   /**
-   * Shuts down the replication manager. Shuts down the individual replica threads and
+   * Shuts down the replication engine. Shuts down the individual replica threads and
    * then persists all the replica tokens
    * @throws ReplicationException
    */
-  @Override
   public void shutdown() throws ReplicationException {
     try {
       // stop all replica threads
